% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW.R, R/PI.R
\docType{class}
\name{Protein-class}
\alias{Protein-class}
\title{Calculate protein sequence molecular weight}
\arguments{
\item{protein_sequence}{A character string representing the protein sequence.}
}
\value{
Molecular weight of the protein sequence in Daltons (Da).

Isoelectric point (pI) of the protein sequence.
}
\description{
This function calculates the molecular weight of a given protein sequence.

This function calculates the isoelectric point (pI) of a given protein sequence.
}
\examples{
protein_sequence <- "MAEGEITTFTALTEKFNLPPGNYKKPKLLYCSNGGHFLRILPDGTVDGTRDRSDQHIQLQLSAESVGEVYIKSTETGQYLAMDTSGLLYGSQTPSEECLFLERLEENHYNTYTSKKHAEKNWFVGLKKNGSCKRGPRTHYGQKAILFLPLPV"
calculate_protein_molecular_weight(protein_sequence)
protein_sequence <- "MAEGEITTFTALTEKFNLPPGNYKKPKLLYCSNGGHFLRILPDGTVDGTRDRSDQHIQLQLSAESVGEVYIKSTETGQYLAMDTSGLLYGSQTPSEECLFLERLEENHYNTYTSKKHAEKNWFVGLKKNGSCKRGPRTHYGQKAILFLPLPV"
calculate_protein_pI(protein_sequence)
}
